requirements.txt
cbpro==1.1.4
certifi==2021.5.30
chardet==4.0.0
cycler==0.10.0
decorator==5.0.9
environs==9.3.2
idna==2.10
influxdb-client==1.17.0
kiwisolver==1.3.1
marshmallow==3.12.1
matplotlib==3.4.2
msgpack==1.0.2
numpy==1.20.3
pandas==1.2.4
Pillow==8.2.0
py==1.10.0
pyarrow==4.0.0
pymongo==3.5.1
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.17.1
pytz==2021.1
ratelimit==2.2.1
requests==2.13.0
retry==0.9.2
Rx==3.2.0
six==1.10.0
sortedcontainers==2.4.0
urllib3==1.26.5
websocket-client==0.40.0
